+++ test/SubUnsubTest.cpp
@@ -11,21 +11,21 @@ int main() {
   MQProcessor::Queue<int, int> mqproc;
   std::vector<TestConsumer> consumers(countCons, TestConsumer());
   std::vector<Providers<BufferSize>> providers;
-  for (int i = 0; i < countCons; i++)
+  for (size_t i = 0; i < countCons; i++)
     mqproc.Subscribe(i, &consumers[i]);
-  for (int i = 0; i < countCons; i++)
+  for (size_t i = 0; i < countCons; i++)
     providers.emplace_back(Providers<BufferSize>(mqproc, countThreads, i));
-  if( !mqproc.Unsubscribe(countCons / 2) )
+  if (!mqproc.Unsubscribe(countCons / 2))
     return 1;
-  if( !mqproc.Unsubscribe(countCons / 4) )
+  if (!mqproc.Unsubscribe(countCons / 4))
     return 2;
-  if( !mqproc.Subscribe(countCons / 2, &consumers[countCons / 2]) )
+  if (!mqproc.Subscribe(countCons / 2, &consumers[countCons / 2]))
     return 3;
-  if( !mqproc.Unsubscribe(countCons / 8) )
+  if (!mqproc.Unsubscribe(countCons / 8))
     return 4;
-  if( mqproc.Unsubscribe(countCons + 1) )
+  if (mqproc.Unsubscribe(countCons + 1))
     return 5;
-  if( mqproc.Subscribe(0, &consumers[0]) )
+  if (mqproc.Subscribe(0, &consumers[0]))
     return 6;
   for (auto &i : providers)
     i.Start();
